post

+++ app/Http/Controllers/TasksController.php
@@ -10,11 +10,17 @@ class TasksController extends Controller
 {
     public function index()
     {
-        // タスク一覧を取得
-        $tasks = Task::paginate(25);
-        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザのタスク一覧を取得(作成日時の降順)
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(25);
+            
+            $data = [ 'user' => $user, 'tasks' => $tasks, ];
+        }
         // タスク一覧ビューでそれを表示
-        return view('tasks.index', [ 'tasks' => $tasks, ]);
+        return view('tasks.index', $data);
     }
 
     public function create()
@@ -33,12 +39,12 @@ class TasksController extends Controller
             'content' => 'required',
         ]);
         
-        // タスクを作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+        // 認証済みユーザ（閲覧者）のタスクとして作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'status' => $request->status, 
+            'content' => $request->content,
+        ]);
+        
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -48,8 +54,14 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // タスク詳細ビューでそれを表示
-        return view('tasks.show', [ 'task' => $task, ]);
+        // 認証済みユーザのタスクである場合は、タスク詳細ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [ 'task' => $task, ]);
+        }
+        // 認証済みユーザのタスクでない場合は、トップページにリダイレクト
+        else{
+            return redirect('/');
+        }
     }
 
     public function edit($id)
@@ -57,8 +69,14 @@ class TasksController extends Controller
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
         
-        // タスク編集ビューでそれを表示
-        return view('tasks.edit', [ 'task' => $task, ]);
+        // 認証済みユーザのタスクである場合は、タスク編集ビューでそれを表示
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [ 'task' => $task, ]);
+        }
+        // 認証済みユーザのタスクでない場合は、トップページにリダイレクト
+        else{
+            return redirect('/');
+        }
     }
 
     public function update(Request $request, $id)
@@ -84,8 +102,10 @@ class TasksController extends Controller
     {
         // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // タスクを削除
-        $task->delete();
+        // 認証済みユーザである場合は、タスクを削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         // トップページへリダイレクトさせる
         return redirect('/');
